+++ shellfoundry/bootstrap.py
@@ -74,11 +74,13 @@ def install():
 
 
 @cli.command()
-def dist():
+@click.option('--enable_cs_repo', is_flag=True, help="Includes shell dependencies "
+                                                     "that are stored in the local pypi repository")
+def dist(enable_cs_repo):
     """ Creates a deployable Shell which can be distributed to a production environment """
 
     PackCommandExecutor().pack()
-    DistCommandExecutor().dist()
+    DistCommandExecutor().dist(enable_cs_repo)
 
 
 @cli.command()