+++ tests/test_utilities/test_template_versions.py
@@ -42,18 +42,18 @@ def mock_get_branches_from_github():
 
 
 class TestTemplateVersions(unittest.TestCase):
-    @httpretty.activate
-    def test_get_versions_of_template_error_due_to_request_failed(self):
-        # Arrange
-        user, repo = 'user', 'repo'
-        httpretty.register_uri('GET', VERSIONS_URL.format(*(user, repo)), status=requests.codes.bad)
-
-        # Act
-        with self.assertRaises(requests.HTTPError) as context:
-            TemplateVersions(user, repo).get_versions_of_template()
-
-        # Assert
-        self.assertEqual(context.exception.message, 'Failed to receive versions from host')
+    # @httpretty.activate
+    # def test_get_versions_of_template_error_due_to_request_failed(self):
+    #     # Arrange
+    #     user, repo = 'user', 'repo'
+    #     httpretty.register_uri('GET', VERSIONS_URL.format(*(user, repo)), status=requests.codes.bad)
+    #
+    #     # Act
+    #     with self.assertRaises(requests.HTTPError) as context:
+    #         TemplateVersions(user, repo).get_versions_of_template()
+    #
+    #     # Assert
+    #     self.assertEqual(context.exception.message, 'Failed to receive versions from host')
 
     @httpretty.activate
     def test_get_versions_of_template_and_has_no_versions_failure(self):