+++ HISTORY.rst
@@ -2,6 +2,16 @@
 History
 =======
 
+1.2.9 (2019-03-27)
+------------------
+
+* Added possibility to download dependencies from local CS repository during 'shellfoundry dist' command
+
+1.2.8 (2019-03-05)
+------------------
+
+* Fixed issue with wrong setuptools import
+
 1.2.7 (2019-02-12)
 ------------------